% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_defaults}
\alias{auto_defaults}
\title{Set auto theming defaults}
\usage{
auto_defaults(bg = NULL, fg = NULL, accent = NULL, font = NULL)
}
\arguments{
\item{bg}{a background color.}

\item{fg}{a foreground color.}

\item{accent}{a color for making certain graphical markers 'stand out'
(e.g., the fitted line color for \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}).
Can be 2 colors for lattice (stroke vs fill accent).}

\item{font}{a \code{\link[=font_spec]{font_spec()}} object. If missing, font defaults are not altered.}
}
\description{
Auto theming defaults are used to resolve \code{"auto"} values outside
of a \strong{shiny} runtime (i.e., where auto theming might be based on
imperfect heuristics). Setting of auto defaults is especially useful
for developers of a custom rmarkdown format that wish to have better
default auto-theming behavior. By having the output document call
\code{auto_defaults()} "pre-knit" (and restoring the old defaults "post-knit"),
users of the output document can then simply call \code{thematic_on()} within
their document to adopt these defaults.
}
\details{
Call this function with no arguments to get the current auto defaults.
}
\examples{
auto_defaults("black", "white")
thematic_on()
plot(1:10, 1:10)
}
