% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{Cat}
\alias{Cat}
\title{Categorical Scale :: generates a palette with n steps || GreenBlue -> RedOrange -> Purple -> Pink -> Green -> Yellow -> GoldBrown -> Black}
\usage{
Cat(n, a = NA, name = "Dark2")
}
\arguments{
\item{n}{the number of steps for color scale}

\item{a}{the alpha/opacity of the colors}

\item{name}{the name of scale found within ColorBrewer, currently defaulted to Dark2}
}
\description{
Categorical Scale :: generates a palette with n steps || GreenBlue -> RedOrange -> Purple -> Pink -> Green -> Yellow -> GoldBrown -> Black
}
