% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabGT.R
\name{cell_format}
\alias{cell_format}
\title{Format cell line}
\usage{
cell_format(
  table_data = NULL,
  cell_line_color = "black",
  cell_weight = 1,
  cell_style = "solid"
)
}
\arguments{
\item{table_data}{data set that will be used to create the table}

\item{cell_line_color}{color of cell line}

\item{cell_weight}{thickness of line}

\item{cell_style}{type of line}
}
\description{
Format cell line
}
