% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{Cer_Cts}
\alias{Cer_Cts}
\title{Continuous Scale :: generates a palette with n steps || White -> Cerulean}
\usage{
Cer_Cts(n, a = NA)
}
\arguments{
\item{n}{the number of steps for color scale}

\item{a}{the alpha/opacity of the colors}
}
\description{
Continuous Scale :: generates a palette with n steps || White -> Cerulean
}
