% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{Cer_Div}
\alias{Cer_Div}
\title{Divergent Scale :: generates a palette with n steps || Gray -> White -> Cerulean}
\usage{
Cer_Div(n, a = NA)
}
\arguments{
\item{n}{the number of steps for color scale}

\item{a}{the alpha/opacity of the colors}
}
\description{
Divergent Scale :: generates a palette with n steps || Gray -> White -> Cerulean
}
