% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabGT.R
\name{column_format}
\alias{column_format}
\title{Formats values within cells by column}
\usage{
column_format(
  table_data = NULL,
  dec = 2,
  trail = TRUE,
  date_col = NULL,
  style = 1
)
}
\arguments{
\item{table_data}{data set that will be used to create the table}

\item{dec}{decimal place of the number data}

\item{trail}{determines if there should be the removal of trailing 0's}
}
\description{
Formats values within cells by column
}
