% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{font_cache_set}
\alias{font_cache_set}
\title{Control the directory used for font caching}
\usage{
font_cache_set(path, cleanup = FALSE)
}
\arguments{
\item{path}{a filepath.}
}
\value{
Returns the previously used caching directory.
}
\description{
Control the directory used for font caching
}
