% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{font_spec}
\alias{font_spec}
\title{Font specification}
\usage{
font_spec(
  families = "",
  scale = 1,
  auto_install = is_installed("ragg") || is_installed("showtext")
)
}
\arguments{
\item{families}{a character vector of font families.}

\item{scale}{numerical constant applied to font sizes.}

\item{auto_install}{whether or not to attempt automatic download and registration
of fonts not found on the system. Currently any font on Google Fonts is supported.}
}
\value{
a list of information about the font specification.
}
\description{
Specify a collection of font families. The first font family supported
by the relevant device (i.e., the device that is open, or will be opened, at
plotting time) is used by thematic. If a given font family is not supported
by the default, but is a \href{https://fonts.google.com/}{Google Font} and
\code{auto_install = TRUE}, the font will be downloaded, cached, and registered
for use the \strong{showtext} and \strong{ragg} packages.
}
\seealso{
\code{\link[=thematic_with_device]{thematic_with_device()}}, \code{\link[=thematic_begin]{thematic_begin()}}, \code{\link[=font_cache_set]{font_cache_set()}}
}
