% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabGT.R
\name{footnote_format}
\alias{footnote_format}
\title{Footnote format}
\usage{
footnote_format(
  table_data = NULL,
  ...,
  footnote_line_color = "black",
  footnote_weight = 2,
  footnote_style = "solid"
)
}
\arguments{
\item{table_data}{data set that will be used to create the table}

\item{...}{values to be placed within the tab_foot() function}

\item{footnote_line_color}{color of border line}

\item{footnote_weight}{thickness of border line}

\item{footnote_style}{style of border line}
}
\description{
Footnote format
}
