% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{gg_tool}
\alias{gg_tool}
\title{Theme tool for generic ggplot}
\usage{
gg_tool(
  complete_themes = theme_bw(),
  title_label = "",
  subtitle_label = "",
  title_caption = "",
  leg.title = "",
  x_label = "",
  y_label = "",
  title.pos = element_text(hjust = 0.5),
  subtitle.pos = element_text(hjust = 0.5),
  leg.pos = "right",
  angle = 30,
  hjust = 1,
  vjust = 1.5
)
}
\arguments{
\item{complete_themes}{plot background theme}

\item{title_label}{plot title}

\item{subtitle_label}{plot subtitle}

\item{title_caption}{plot caption}

\item{x_label}{x axis label}

\item{y_label}{y axis label}

\item{...}{theme information}
}
\description{
Theme tool for generic ggplot
}
