% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{GYR}
\alias{GYR}
\title{G-Y-R Scale :: generates a palette with 3 steps || Green-> Yellow-> Red}
\usage{
GYR(n = 3, a = NA)
}
\arguments{
\item{n}{the number of steps for color scale}

\item{a}{the alpha/opacity of the colors}
}
\description{
G-Y-R Scale :: generates a palette with 3 steps || Green-> Yellow-> Red
}
