% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{Mag}
\alias{Mag}
\title{Magma Scale :: generates a palette with n steps || Black -> Purple -> Red -> Orange -> Yellow}
\usage{
Mag(n, a = NA)
}
\arguments{
\item{n}{the number of steps for color scale}

\item{a}{the alpha/opacity of the colors}
}
\description{
Magma Scale :: generates a palette with n steps || Black -> Purple -> Red -> Orange -> Yellow
}
