% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{sequential_gradient}
\alias{sequential_gradient}
\title{Construct a sequential colorscale from fg, accent, and bg}
\usage{
sequential_gradient(alpha = 0.5, n = 30, fg, accent, bg)
}
\arguments{
\item{alpha}{a number between 0 and 1. Controls how close the endpoints
should be to the \code{fg} and \code{bg}.}

\item{n}{the number of color codes to return}

\item{fg}{a foreground color.}

\item{accent}{a color for making certain graphical markers 'stand out'
(e.g., the fitted line color for \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}).
Can be 2 colors for lattice (stroke vs fill accent).}

\item{bg}{a background color.}
}
\value{
a vector of color codes.
}
\description{
Construct a sequential colorscale from fg, accent, and bg
}
\seealso{
\code{\link[=thematic_begin]{thematic_begin()}}
}
