% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabDT.R
\name{tabDT}
\alias{tabDT}
\title{Personalized DT Table}
\usage{
tabDT(
  dat = NULL,
  identity = "ID",
  identity_col = c(ID = 1),
  cell_colors = "white",
  table_number = "Table: ",
  table_title = "title",
  btn = c("csv", "excel", "pdf")
)
}
\arguments{
\item{dat}{data set that will be used to create the table}

\item{identity}{name of row count}

\item{identity_col}{the location and name of id column}

\item{cell_colors}{color of cells}

\item{table_number}{number of the given table}

\item{table_title}{title of table}

\item{btn}{selection of buttons csv, excel, and pdf}
}
\description{
Personalized DT Table
}
