% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabGT.R
\name{tabGT}
\alias{tabGT}
\title{Formats data to static table}
\usage{
tabGT(
  table_data = NULL,
  table_title = "",
  table_subtitle = "",
  source_doc = "",
  dec = 2,
  trail = TRUE,
  date_col = NULL,
  style = 1,
  cell_line_color = "black",
  cell_weight = 1,
  cell_style = "solid",
  font = "Perpetua",
  font_color = "black",
  main_bgc = "#0197F6",
  secondary_bgc = "#91D4FE",
  main_border_color = "black"
)
}
\arguments{
\item{table_data}{data set that will be used to create the table}

\item{table_title}{title in MD}

\item{table_subtitle}{subtitle in MD}

\item{source_doc}{source documentation in MD}

\item{dec}{decimal place}

\item{trail}{selection T/F to allow trailing 0's}

\item{cell_line_color}{color of cell line}

\item{cell_weight}{cell line thickness}

\item{cell_style}{cell line style}

\item{font}{font type}

\item{font_color}{font color}

\item{main_bgc}{background fill color (title/subtitle)}

\item{secondary_bgc}{other fill color (not title or cell area)}

\item{main_border_color}{border color}
}
\description{
Formats data to static table
}
