% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabGT.R
\name{table_option}
\alias{table_option}
\title{Formats font, color for fill and main border}
\usage{
table_option(
  table_data = NULL,
  font = "Bookman",
  font_color = "black",
  main_bgc = "#0197F6",
  secondary_bgc = "#91D4FE",
  main_border_color = "black"
)
}
\arguments{
\item{table_data}{data set that will be used to create the table}

\item{font}{font type}

\item{font_color}{font color}

\item{main_bgc}{background color of title/subtitle}

\item{secondary_bgc}{background color of non-title or cells}

\item{main_border_color}{color of tables border}
}
\description{
Formats font, color for fill and main border
}
