% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic-package.R
\docType{package}
\name{thematic-package}
\alias{thematic}
\alias{thematic-package}
\title{thematic: Unified and Automatic 'Theming' of 'ggplot2', 'lattice', and 'base' R Graphics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Theme 'ggplot2', 'lattice', and 'base' graphics based on a few choices, including foreground color, background color, accent color, and font family. Fonts that aren't available on the system, but are available via download on 'Google Fonts', can be automatically downloaded, cached, and registered for use with the 'showtext' and 'ragg' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/thematic/}
  \item \url{https://github.com/rstudio/thematic}
  \item Report bugs at \url{https://github.com/rstudio/thematic/issues}
}

}
\author{
\strong{Maintainer}: Carson Sievert \email{carson@posit.co} (\href{https://orcid.org/0000-0002-4958-2844}{ORCID})

Authors:
\itemize{
  \item Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Joe Cheng \email{joe@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
