% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{thematic_get}
\alias{thematic_get}
\alias{thematic_get_option}
\alias{thematic_get_mixture}
\title{Query the current thematic theme}
\usage{
thematic_get()

thematic_get_option(name = "", default = NULL)

thematic_get_mixture(amounts = 0.5)
}
\arguments{
\item{name}{a theme element name (e.g., \code{fg}, \code{bg}, etc.)}

\item{default}{a default value to return in the event no thematic theme is active.}

\item{amounts}{value(s) between 0 and 1 specifying how much to mix \code{fg} (0) and \code{bg} (1).}
}
\description{
\code{\link[=thematic_get]{thematic_get()}} returns information about the entire theme,
whereas \code{\link[=thematic_get_option]{thematic_get_option()}} returns information about a specific
option.
}
\examples{
thematic_get()
thematic_on("darkblue", "skyblue")
thematic_get_option("bg")

if (interactive()) {
  scales::show_col(thematic_get_mixture(seq(0, 1, by = 0.1)))
}

}
