% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{thematic_get_theme}
\alias{thematic_get_theme}
\alias{thematic_set_theme}
\alias{thematic_with_theme}
\alias{thematic_get_option}
\alias{thematic_get_mixture}
\title{(Temporarily) setting and restoring themes}
\usage{
thematic_get_theme()

thematic_set_theme(theme)

thematic_with_theme(expr, ...)

thematic_get_option(name = "", default = NULL)

thematic_get_mixture(amounts = 0.5)
}
\arguments{
\item{theme}{a \code{theme} object (as returned by \link{thematic_on} or \code{\link[=thematic_get_theme]{thematic_get_theme()}})
or \code{NULL} (in which case \code{thematic_off()} is called).}

\item{expr}{R code that produces a plot.}

\item{...}{arguments passed along to \code{\link[=thematic_on]{thematic_on()}}.}

\item{name}{a theme element name (e.g., \code{fg}, \code{bg}, etc.)}

\item{default}{a default value to return in the event no thematic theme is active.}

\item{amounts}{value(s) between 0 and 1 specifying how much to mix \code{fg} (0) and \code{bg} (1).}
}
\description{
Use these functions if you'd like to enable thematic for a single plot
and/or want to program with thematic (i.e. add additional customization in
a custom function):
}
\details{
\itemize{
\item \code{\link[=thematic_with_theme]{thematic_with_theme()}}: similar to \code{\link[=thematic_on]{thematic_on()}}, but for an single plot.
\item \code{\link[=thematic_set_theme]{thematic_set_theme()}}: set a given \code{theme} object as the current theme.
\item \code{\link[=thematic_get_theme]{thematic_get_theme()}}: return a \code{theme} object describing the
current thematic theme.
\item \code{\link[=thematic_get_theme]{thematic_get_theme()}}: get a particular \code{theme} option (with a sensible default
if thematic is not enabled).
\item \code{\link[=thematic_get_mixture]{thematic_get_mixture()}}: get a mixture of \code{bg} and \code{fg}.
}
}
\examples{

# Use thematic_with_theme() for a one-time use of thematic
thematic_with_theme(
  plot(1:10, col = thematic_get_option("accent"), pch = 19),
  "darkblue", "skyblue", accent = "red"
)

# Use thematic_set_theme() if doing something more complicated
# like programming on top thematic (without causing side effects)
my_plot <- function(expr, las = 3, ...) {
  old_theme <- thematic_on("black", "white")
  on.exit(thematic_set_theme(old_theme), add = TRUE)
  opts <- par(las = las)
  on.exit(par(opts), add = TRUE)
  # Imagine some more customization with ...
  force(expr)
}
my_plot(plot(1:10))

# Use thematic_get_option() if you want just part of the theme
# object (and return a default value if thematic isn't enabled)
thematic_off()
thematic_get_option("bg", "white")
thematic_on(bg = "red")
thematic_get_option("bg", "white")
thematic_off()

# Use thematic_get_mixture() if you want to get a mixture of the
# background and foreground colors
thematic_with_theme(
  scales::show_col(thematic_get_mixture(seq(0, 1, by = 0.1))),
  "darkblue", "skyblue"
)

}
