% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabGT.R
\name{title_note_format}
\alias{title_note_format}
\title{Formats table with title, subtitle, and source documentation}
\usage{
title_note_format(
  table_data = NULL,
  table_title = NULL,
  table_subtitle = NULL,
  source_doc = ""
)
}
\arguments{
\item{table_data}{data set that will be used to create the table}

\item{table_title}{title, to be wrote in MD format}

\item{table_subtitle}{subtitle, to be wrote in MD format}

\item{source_doc}{documentation, to be wrote in MD format}
}
\description{
Formats table with title, subtitle, and source documentation
}
