% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{Vir}
\alias{Vir}
\title{Viridis Scale :: generates a palette with n steps || Purple -> Blue -> Green -> Yellow}
\usage{
Vir(n, a = NA)
}
\arguments{
\item{n}{the number of steps for color scale}

\item{a}{the alpha/opacity of the colors}
}
\description{
Viridis Scale :: generates a palette with n steps || Purple -> Blue -> Green -> Yellow
}
