% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{auto_defaults_set}
\alias{auto_defaults_set}
\alias{auto_defaults_get}
\title{Set auto theming defaults}
\usage{
auto_defaults_set(bg = NULL, fg = NULL, accent = NULL, font = NULL)

auto_defaults_get()
}
\arguments{
\item{bg}{a background color.}

\item{fg}{a foreground color.}

\item{accent}{a color for making certain graphical markers 'stand out'
(e.g., the fitted line color for \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}).
Can be 2 colors for lattice (stroke vs fill accent).}

\item{font}{a \code{font_spec()} object. If missing, font defaults are not altered.}
}
\description{
Setting of auto-theming defaults is primarily useful for developers
of a custom rmarkdown format that wish to have better default auto-theming
behavior. By having the output document call \code{auto_defaults_set()}
"pre-knit" (and restoring the old defaults "post-knit"), users of the
output document can then simply call \code{thematic_on()} within their document
to adopt these defaults.
}
\examples{
auto_defaults_set("black", "white")
thematic_on()
plot(1:10, 1:10)
}
