% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{okabe_ito}
\alias{okabe_ito}
\title{Okabe Ito colorscale}
\usage{
okabe_ito(n = NULL)
}
\arguments{
\item{n}{number of colors.}
}
\description{
Okabe Ito colorscale
}
\references{
\url{https://jfly.uni-koeln.de/color/}
}
