% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{okabe_ito}
\alias{okabe_ito}
\title{A color-blind safe qualitative colorscale (Okabe-Ito)}
\usage{
okabe_ito(n = NULL)
}
\arguments{
\item{n}{number of colors.}
}
\value{
a vector of color codes.
}
\description{
This is the default \code{qualitative} colorscale in \code{thematic_on()}
}
\references{
\url{https://jfly.uni-koeln.de/color/}
}
\seealso{
\code{\link[=thematic_on]{thematic_on()}}
}
