% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{thematic_begin}
\alias{thematic_begin}
\alias{thematic_end}
\title{Theme static plots based on a few colors}
\usage{
thematic_begin(
  bg = NULL,
  fg = NULL,
  accent = NA,
  font = NA,
  sequential = sequential_gradient(),
  qualitative = okabe_ito()
)

thematic_end()
}
\arguments{
\item{bg}{a background color.}

\item{fg}{a foreground color.}

\item{accent}{a color for making certain graphical markers 'stand out'
(e.g., the fitted line color for \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}).
Can be 2 colors for lattice (stroke vs fill accent).}

\item{font}{a \code{font_spec()} object. If missing, font defaults are not altered.}

\item{sequential}{a color palette for graphical markers that encode
numeric values. Can be a vector of color codes or a
\code{\link[=sequential_gradient]{sequential_gradient()}} object.}

\item{qualitative}{a color palette for graphical markers that encode
qualitative values (won't be used in ggplot2 when the number of data
levels exceeds the max allowed colors). Defaults to the Okabe-Ito colorscale.}
}
\value{
Returns any information about the previously set theme (if any), invisibly.
}
\description{
Theme ggplot2, lattice, and base graphics based on just a few colors
supplied to \code{\link[=thematic_begin]{thematic_begin()}}. \code{\link[=thematic_begin]{thematic_begin()}} works by modifying global
state (e.g., sets relevant options in \code{\link[graphics:par]{graphics::par()}}, \code{\link[grid:gpar]{grid::gpar()}},
\code{\link[lattice:trellis.par.set]{lattice::trellis.par.set()}}, \code{\link[ggplot2:theme_set]{ggplot2::theme_set()}}, etc). To restore
global state to the state before \code{\link[=thematic_begin]{thematic_begin()}} was called,
use \code{\link[=thematic_end]{thematic_end()}}.
}
\details{
Colors may be anything understood by \code{\link[=col2rgb]{col2rgb()}} or \code{htmltools::parseCssColors()}
(i.e., may be any valid R or CSS color string).
}
\examples{
# simple dark mode
thematic_begin("black", "white")
plot(1:10)
plot(1:10, col = 1:10)
lattice::show.settings()

# use any color code
thematic_begin("#444444", "#e4e4e4")
plot(1:10)
plot(1:10, col = 1:10)
lattice::show.settings()

# restores _original_ state
thematic_end()
plot(1:10)
lattice::show.settings()

thematic_begin("darkblue", "skyblue", "orange")
image(volcano)
image(volcano, col = thematic_get_option("sequential"))
lattice::show.settings()
thematic_end()

}
\seealso{
\code{\link[=font_spec]{font_spec()}}, \code{\link[=thematic_with_device]{thematic_with_device()}}, \code{\link[=thematic_get]{thematic_get()}}
}
