% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{thematic_on}
\alias{thematic_on}
\alias{thematic_off}
\title{Automatic and consistent theming of static plots}
\usage{
thematic_on(
  bg = "auto",
  fg = "auto",
  accent = "auto",
  font = NA,
  sequential = sequential_gradient(),
  qualitative = okabe_ito()
)

thematic_off()
}
\arguments{
\item{bg}{a background color.}

\item{fg}{a foreground color.}

\item{accent}{a color for making certain graphical markers 'stand out'
(e.g., the fitted line color for \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}}).
Can be 2 colors for lattice (stroke vs fill accent).}

\item{font}{a \code{font_spec()} object. If missing, font defaults are not altered.}

\item{sequential}{a color palette for graphical markers that encode
numeric values. Can be a vector of color codes or a
\code{\link[=sequential_gradient]{sequential_gradient()}} object.}

\item{qualitative}{a color palette for graphical markers that encode
qualitative values (won't be used in ggplot2 when the number of data
levels exceeds the max allowed colors). Defaults to the Okabe-Ito colorscale.}
}
\value{
Returns any information about the previously set theme (if any), invisibly.
}
\description{
Turn on (or off) automatic theming of ggplot2, lattice, and base graphics.
While on, thematic registers \code{\link[=plot.new]{plot.new()}}/\code{\link[=grid.newpage]{grid.newpage()}} hooks
that set relevant options (i.e., \code{\link[graphics:par]{graphics::par()}}, \code{\link[grid:gpar]{grid::gpar()}},
\code{\link[lattice:trellis.par.set]{lattice::trellis.par.set()}}, \code{\link[ggplot2:theme_set]{ggplot2::theme_set()}}, etc) based on
the current context.
}
\details{
The \code{bg}, \code{fg}, \code{accent}, and \code{font} arguments all support a value of \code{'auto'},
which are all resolved (just before plotting) based on the following (global)
information, where available
\enumerate{
\item \code{shiny::getCurrentOutputInfo()}.
\item \code{\link[=auto_preferences_set]{auto_preferences_set()}}.
\item \code{bootstraplib::bs_theme_get_variables()}.
\item \code{rstudioapi::getThemeInfo()}.
}

Note that if the resolving of \code{'auto'} for some reason doesn't pick up the
right information, you can always specify values in \code{thematic_on()}.
Colors (e.g., \code{bg}, \code{fg}, \code{accent}) may be any value understood by \code{\link[=col2rgb]{col2rgb()}}
or \code{htmltools::parseCssColors()} (i.e., may be any valid R or CSS color string).
}
\examples{
# simple dark mode
thematic_on("black", "white")
plot(1:10)
plot(1:10, col = 1:10)
lattice::show.settings()

# use any color code
thematic_on("#444444", "#e4e4e4")
plot(1:10)
plot(1:10, col = 1:10)
lattice::show.settings()

# restores _original_ state
thematic_off()
plot(1:10)
lattice::show.settings()

thematic_on("darkblue", "skyblue", "orange")
image(volcano)
image(volcano, col = thematic_get_option("sequential"))
lattice::show.settings()
thematic_off()

}
\seealso{
\code{\link[=font_spec]{font_spec()}}, \code{\link[=thematic_with_device]{thematic_with_device()}}, \code{\link[=thematic_get]{thematic_get()}}
}
