% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic-save-plot.R
\name{thematic_with_device}
\alias{thematic_with_device}
\title{Capture a plot with a graphics device}
\usage{
thematic_with_device(...)
}
\description{
This function is deprecated. Use \code{\link[=thematic_save_plot]{thematic_save_plot()}} instead.
}
\keyword{internal}
